feat: widen BusDebug/MachineDebug debug addresses u16 -> u32
- BusDebug::read/write and the set/clear_watchpoint defaults (plus the
  MachineDebug delegating pair) take addr: u32, per the Address Width
  section of docs/designs/debug-observability.md
- #[derive(BusDebug)] generates u32 signatures that shim back to u16
  with a try_from guard, so 16-bit boards' maps and explicit
  read/write methods stay untouched and addresses above 0xFFFF read
  as unmapped / are ignored
- Galaga and Dig Dug manual impls widen the same way; their
  board-level Watchpoints calls drop the now-redundant 'as u32'
- Default BusDebug::peek loses its 0xFFFF guard (read is u32 now) and
  its test bus grows a high window proving >0xFFFF reaches impls
- debug_ui call sites widen losslessly; UI state stays u16 until the
  32-bit UI subtask

+++ core/src/core/machine.rs
@@ -200,7 +200,7 @@ pub trait MachineDebug {
     /// Set a memory watchpoint in the address space of `cpu_index`.
     ///
     /// Default: delegates to `BusDebug::set_watchpoint()` via `debug_bus_mut()`.
-    fn set_watchpoint(&mut self, cpu_index: usize, addr: u16, kind: WatchpointKind) {
+    fn set_watchpoint(&mut self, cpu_index: usize, addr: u32, kind: WatchpointKind) {
         if let Some(bus) = self.debug_bus_mut() {
             bus.set_watchpoint(cpu_index, addr, kind);
         }
@@ -209,7 +209,7 @@ pub trait MachineDebug {
     /// Clear a memory watchpoint in the address space of `cpu_index`.
     ///
     /// Default: delegates to `BusDebug::clear_watchpoint()` via `debug_bus_mut()`.
-    fn clear_watchpoint(&mut self, cpu_index: usize, addr: u16, kind: WatchpointKind) {
+    fn clear_watchpoint(&mut self, cpu_index: usize, addr: u32, kind: WatchpointKind) {
         if let Some(bus) = self.debug_bus_mut() {
             bus.clear_watchpoint(cpu_index, addr, kind);
         }
